requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-multipart
numpy
scipy
imageio
pillow
pyotp
qrcode
